Lock down minitest to version 4
The tests are using both an around method in the base class and a before
block in the test itself. In minitest 4, the before block runs first,
then the around method. In minitest 5, the around runs first, with the
before block being called at the yield. The tests in mr-sparkle depend
on setting a variable in a before block and using it in the around
block, so it breaks in minitest 5.

The temporary here is to lock it down to minitest 4 in the gemspec.  In
the future, tests should be redone not to depend on this.